requirements.txt
python-telegram-bot==20.7
anthropic
httpx==0.25.2
python-dotenv
pytz
